review changes: remove ccvbars, rename components that are not generic ui components, added missing icon for secondary site backup, min table column width
